feat: add ft_handle_input
