+++ src/main.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gwolf <gwolf@student.42vienna.com>         +#+  +:+       +#+        */
+/*   By: sqiu <sqiu@student.42vienna.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/12 15:15:13 by gwolf             #+#    #+#             */
-/*   Updated: 2023/07/21 16:53:59 by gwolf            ###   ########.fr       */
+/*   Updated: 2023/07/21 17:31:06 by sqiu             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,6 @@
  */
 
 #include "minishell.h"
-#include "mod_lexer.h"
-/* #include "lexer_list.h"
-#include "lexer_tok_utils.h" */
 
 #ifdef TESTING
 # define main not_main
@@ -28,14 +25,12 @@ int	main(int argc, char **argv)
 {
 	t_data		*data;
 	char 		*input;
-	t_tkn_list	*lst;
 
 	(void)argc;
 	(void)argv;
 	data = malloc(sizeof(*data));
 	if (!data)
 		ft_exit_failure(data, ERR_MALLOC);
-	lst = NULL;
 	if (ft_env_setup(&data->env_table) != SUCCESS)
 		printf("NO\n");
 	ft_hashtable_insert(data->env_table, "PS1=\\u@\\h:\\w$ ", 3);
@@ -51,7 +46,9 @@ int	main(int argc, char **argv)
 		if (!input)
 			break ;
 		add_history(input);
-		ft_lex_input(&lst, input);
+		data->err = ft_handle_input(input);
+		if (data->err != SUCCESS)
+			ft_exit_failure(data, data->err);
 		//do stuff
 		free(input);
 		free(data->prompt1);
